Revert "refactor: rename abstract to generic"
This reverts commit 89d30eb23d90f2ff27ec92cf47c30f0e8d9ac347.
