+++ cloudevents/http/event.py
@@ -18,10 +18,10 @@ import uuid
 
 import cloudevents.exceptions as cloud_exceptions
 from cloudevents.http.mappings import _required_by_version
-from cloudevents import generic
+from cloudevents import abstract
 
 
-class CloudEvent(generic.CloudEvent):
+class CloudEvent(abstract.CloudEvent):
     """
     Python-friendly cloudevent class supporting v1 events
     Supports both binary and structured mode CloudEvents
